+++ src/components/Card/Post/Detail/DetailPost.tsx
